[editcontroller] backout (sort of) removal of entity.complete() in validate_form
entity is passed to js callbacks when there's one. In that case, we want as much
information as we can.

The removal was there to avoid complete to fail when all constraints (required
relations / attributes) aren't satisfied. This fix only does the complete() after the
commit is done, any ValidationError should have been raised at this point.

--HG--
branch : stable