[web] set pageid at request instanciation rather than in htmlheader template
In some occasions, we need to force a pageid: typically an ajax-rendererd
creation form would need to use the pageid of the main page to correclty
set it's page_data (e.g. rql varmker).

If pageid is found in req.form, it's used, otherwise we build a new one with.
make_uid(). The behaviour should be consistent with the former one in
JSonController.

--HG--
branch : stable