[web] get rid of spurious '[Errno 2] No such file or directory' on load
This was caused by can_do_pdf_conversion() which tries to use
'/usr/bin/fop'. When fop was not available, an OSError was raised
by suprocess.Popen.

This changeset first checks for /usr/bin/fop presence, then tries
to run fop. If an error occurs, log it instead of printing it.

NOTE for later: why do we test explicitly for "/usr/bin/fop"
                rather than simply "fop" ?

--HG--
branch : stable