[storages] fix fs_importing side-effect on entity.data
When creating a new File object, if fs_importing is set,
we want entity.data to be the file content instead of the
filepath for the rest of the transaction.
(see test_bfss_fs_importing_transparency) for test implementation

To make this possible, the storage hooks (entity_added / entity_updated)
must return the correct value to set in the entity dict.

--HG--
branch : stable