[fix] fix path unicity process in BytesFileSystemStorage.new_fs_path
The previous implementation was bugged (prefixing the whole path with '_'
instead of the base name). A new version (using number) replace it.
* * *
Improve BytesFileSystemStorage.new_fs_path to use available metadata

This version try to get some hint about how to name the file using metadata.
Using the real file name and extension when available.  Keeping the extension
might be usefull for exemple in the case of processing that use filename
extension to detect content-type.

--HG--
branch : stable