[devtools] Firefox need write access to $HOME
Check that you have write acces to the user home before launching Firefox.
Firefox need to write stuff in the home directory of the user and will silency
crash when unable to do it. We add an assertion with an explicite message to
improve crash report in such situation.

--HG--
branch : stable