[devtools] make a copy of the xvfb-run script so we can kill it properly
Since changeset a3d3bdd46463 we don't run firefox directly but through a
wrapper script.  This meant that on stopping the test, we would kill
that wrapper script but leave the Xvfb server and firefox processes
running.  So make a local copy of xvfb-run, and make it kill Xvfb when
it goes down, not just when its client exits.

--HG--
branch : stable