[entity/hooks] fix regression introduced by 4ba11607d84a
on create_entity, attributes weren't cached, but they are since 4ba11607d84a
and this triggered a bug: used values may be transformed by hooks (e.g.
TidyHtmlFields), ending in invalid value in the cache. This is the regression.

Now, set_attributes used to update the attributes cache while suffering from
the very same pb. This is the now fixed old bug.

Transformed attributes are now recorded in a transaction shared variable, and
those are omitted from the client cache.