[composite] apply composite on new enty too (closes #2450679)
Before this changesets, automatic deletion did not applied to entity created in
the same transaction. This patch remove this filtering in the dedicated
operation.

The original filtering where introduced by 5d889b4928bb but no rational where
found for this changes.

The former behavior is seen as inconsistent and creates bug in some cubes.

--HG--
branch : oldstable