[entity attr cache] mark attribute as uncacheable in the underlying function else we may miss some changes.
Also rewrite a storage test currently failing because cache of the entity created by the
test transaction, distinct from the entity created internally and given to hooks and all,
has its attributes cache not updated. As this doesn't seems a proper usage, rewrite it
as expected.

Much probably closes #2423719 definitly.