[service] drop the asynchronous execution possibility
Call_service was able of both sync and sync execution making the API confusing.
There is not user of the async case. We drop the async argument in favor of
synchronous execution only.

This makes call_service the official API to call server side code from the
client side. This is a remplacement for the usual monkey patching of the repo
object.

The zmq notification bus is a solid alternative for codes that needs to start an
async execution.