[repoapi] introduce a basic ClientConnection class
This is the new official way to access the repo from client side.

It still access Session object directly as the server side connection is not up
yet (and it's not up because it would have no user).

Multiple follow up commit will install compatibility with the DBAPI. This will
ease the migration from dbapi to repoapi.

ClientConnection has no user yet but later commit will use it in the whole Web
stack.

related to #2503918