[constraint] more robust unicity constraint failures reporting for end-users
Postgres or Sqlserver have limits on the index names (around resp. 64
and 128 characters). Because `logilab.database` encodes the `unique
together` constraint rtypes in the index names, we sometimes get
truncated index names, from which it is impossible to retrieve all
rtypes.

In the long run, the way such index are named should be changed.

In the short term, we try to reduce the end-user confusion resulting
from this design flaw:

* in source/native, the regex filtering ``IntegrityError`` message does
  not impose an `_idx` suffix, which indeed may be absent (the result being an
  UI message that resembles a catastrophic failure),

* also we avoid including a trailing " (double quote) from the error
  message

* in entities/adapters, the well-named ``IUserFriendly`` adapter is made a
  bit smarter about how to handle missing rtypes.

* the adapter also always produces a global message explaining the
  issue (and the fact that sometimes, the user is not shown all the
  relevant info)

* i18n is updated

Closes #2793789

--HG--
branch : stable