[web] don't rename resources in gen-static-datadir
We should use the result from WebConfiguration.locate_resource to know
which file to copy, not what name to copy it to.  The destination file
should be the name of the resource, as the client will request it.

This matters in the case of cubicweb.css which locate_resource might
rewrite to cubicweb.old.css.

Closes #3206129

--HG--
branch : stable