[req] New method: RequestSessionBase.find().
This method does what find_entities and find_one_entity did, except it
returns the resultset itself.

In addition, it accepts 'reverse_' arguments and check that the relations
actually exists on the entity before executing the query.

Also, reimplement find_one_entity and find_entity based on the new function
so they benefit from the more complete implementation, and deprecate them.

Note: List of values in kwargs are NOT supported in this initial
implementation.

Closes #3361290