[rql2sql] remove special behaviour of symmetric relation vs DISTINCT
0542a85fe667 replacing SQL OR by hooks for symmetric relations allows that.

This involve a single test value change for a border case: when querying a
symmetric relation without specifying the subject nor the object, you may get
some duplicated result. IMO this is fine to let the user explicitly use DISTINCT
or not and to remove the dedicated handling we had which didn't let any choice.

Related to #3259713