[migration] always rebuild infered relation
This was skipped for some bad reason (see 12ad88615a12 which
introduced the change).

Fix for #231956 in Yams is necessary to allow this cset: during a
migration, we want to always reinfer relations while allowing explicit
redefinition of an infered relation later. Else, we may run the
migration with missing parts of the schema (the one that should have
been infered).

Closes #3685463.

--HG--
branch : stable