Skip core_handle, add a context manager to handle cubicweb errors
The context manager is also used to catch errors in render_view.
It handles the 'external' errors raised by cubicweb code.

The more internal errors, the one that should occur only in url resolving and
cubicweb controllers, are handled directly in CubicWebPyramidHandler.

ValidationError is handled by CubicWebPyramidHandler for now, but should
probably be handled by cw_to_pyramid

Related to #4291173