[web] set Vary response header to "Accept-Language" when using content negotiation
This is slightly annoying because the response actually only varies
based on the language we decide to send, which has much fewer possible
values than Accept-Language, but that's not in the request, so we can't
easily use it.  Deployments using varnish or similar and controlling the
set of available languages will likely want to override this to allow
reasonable amounts of caching.

Closes #2105812