Correctly pass the multiple parameters to the cubicweb request
When naively converted to a dict, the webob MultiDict will not set the multiple
parameters (a same name with several values) in a way CW can understand.

MultiDict.mixed() however, does exactly what CW needs: list values for keys
with multiple values, and single value for the others.

Related to #4291181