[server] change order of entities table modification vs entity creation/deletion
In order to have a foreign key referencing the entities table, the
insertion into entities must happen first, and the deletion must happen
last.

For the deletion case, this means we need to:
1. delete all relations (cascade)
2. delete the entities themselves
3. delete the corresponding lines from the entities table
This means the _delete_info{,_multi} methods can't keep doing 1 and 3.

Thankfully the "public" delete_info method appears to be unused, so drop
it.

Related to #4846892.