[uilib] `uilib.js` helper now honors explicit JSString (closes #4959538)
Previously, if a JSString was inside a nested container (i.e. list or
dict), it was handled as a string, not a verbatim js chunk (see
docstring and test). So the following code::

  >>> str(js.cw.pouet(1, {'callback': JSString('cw.mycallback')})

resulted in::

  "cw.pouet(1, {'callback': 'cw.mycallback')})"

while we expect (note the removed quotes)::

  "cw.pouet(1, {'callback': cw.mycallback)})"


We use ``cubiweb.utils.js_dumps`` instead of
``cubicweb.utils.json_dumps`` which is aware of JSString.