[predicates] ExpectedValuePredicate now accepts a dict parameter
``match_form_params`` and ``match_kwargs`` benefit from that.

For instance, the following statements are valid::

  __select__ = match_form_params('vid', 'subvid')
  __select__ = match_form_params(vid='list', subvid='incontext')
  __select__ = match_form_params(vid=('list', 'tsearch'))

In the latter cases, not only the parameters (``vid``/``subvid``)
must be in the request form but their corresponding value must also match
the expected values (or one of possible choices if `values` is a sequence).


closes #5484070