[tox] Remove set of PATH environment variable
The : concatenation does not work actually so the resulting PATH was just
wrong (it did not contain the previous PATH). Anyways, the PATH environment
variable is passed by default by tox so it's probably better to let the caller
set it (thus avoiding to hard-code the specific postgres version and path).

http://tox.readthedocs.org/en/latest/example/basic.html#passing-down-environment-variables