[data] Use correct syntax to unset CSS attributes (closes #8602446)
'default' is actually a recognized CSS cursor name (the regular
pointer). Using this value doesn't revert to whatever cursor DOM
elements have by default ('pointer' on links, 'text' on text, etc).

The proper way to unset a CSS attribute on DOM elements is to set it to
an empty string.

http://stackoverflow.com/questions/2027935/how-to-remove-css-property-using-javascript