[server] always monkeypatch QuerierHelper to handle dates and times on SQLite
The patch prevents problems due to a SQLite bug causing some values to
be returned as strings instead of properly typed objects.

This should probably be improved to use SQLite connection hooks.  But
the major improvement is that the monkeypatch is now done all the time,
and not just for tests.