[devtools] fix get_bounds
Say an attribute has two IntervalBoundConstraints, one with minvalue=0
and the other with maxvalue=10 (not literally, or that could be a single
constraint, but e.g. with references to other attributes).  We don't want to
override the minvalue of the former with None just because the latter happens
last in the constraints list.

This doesn't fix the case where two constraints both have a min or a max
value.  Meh.