[autoform] Avoid two calls to field.process_form for the same field in some cases
when some entity is being created and data include non-inlined relations, the
values for this relation are now stored for later usage, avoiding two calls
to field's process_form method, which may be unexpected for custom fields.

This has been discovered in saem_ref#f5444b1f9770.


Reorganize imports in the test along the way.