[sql gen] Control size of index name using an md5 hash
When using concatenation of table and column, we may hit the backend size limit
on index names, which results on truncated name (64 for postgres). Then we may
have conflict or be unable to find expected index on later synchronization
attempt.

This is done by monkeypatching lgdb for now to avoid dependancy problems.

Unique and primary key indices of entity types handling is left to the backend,
as they are not explicitly named.

See https://www.logilab.org/ticket/4789189

Closes #13821949