[web] Fix bug with usage of os.rename under windows environment
In 7c386161ebd6 we removed cache handling from property sheet and introduced
usage of a tempfile + os.rename to get atomic generation of files. The pb is
that this is not portable, since under windows os.rename will raise an exception
if the file already exists (because there is no way to write a file atomatically
in such case).

This kind of thing should be out of the CW scope anyway, so implements a quick &
dirty fix in the mean time.

Closes #14214794

--HG--
branch : 3.22