[pkg] Declare extras (optional) dependencies
This is primary to extract "pyramid" dependencies out of install_requires
as they are actually optional. Along the way, resurrect all things in
__pkginfo__.__recommends__ which were not used at all and convert them
into extra_requires.

It appears that __depends__ and __recommends__ in __pkginfo__.py are
not useful so drop them and inline dependencies as
install_requires/extra_requires in setup function call.

For pyramid tests to continue working, add respective dependencies to
test-misc.txt requirements file.