[migration] Fix addition of entity type including boundary constraints on its own attributes
This was failing because of the sequence of sql executing when adding an
attribute:

1. add entity type <commit>
2. add relation type for attribute 1 <commit>
3. add attribute 1 and associated constraints <commit>
etc.

In the case of e.g. start/end constraint, we were trying to add the constraint
before addition of the constrained attribute (e.g. add constraint on 'start'
referencing 'end', but only 'start' has been added yet, not 'end').

This patch fix this by:

* adding the relation type to the schema without having to commit, but keeping
  the operation to revert the addition if necessary - this allows to a single
  commit for all attributes of the entity type ;

* using a LateOperation on constraint operation, so we ensure attributes are
  actually added before any constraint is added.

--HG--
branch : 3.24