[rqlrewrite/schema] Avoid parsing computed relations for each query
The RQLRelationRewriter is instanciated for each RQL query, it should
avoid parsing computed relations formula by using a cache, which seems rightly
located on the instance's schema.

This brings a *huge* performance boost to some pages on application with a few
computed relations (x4 observed on a client app).

Kudos to Adrien, David and Sylvain.

Closes #17059828

--HG--
branch : 3.25