[tox] Do not use site packages in testenv
This is the normal way of running tox. We used to set this because
some people liked to run tests using packages installed by other
means (i.e. system-wide). Now, there seems to be a consensus that
running tox without site packages is a better idea, in particular
since many of our dependencies are now available as wheels and thus
would install quickly provided one uses a recent pip/virtualenv.