[rqlrewrite] Fix rewrite on ambiguities introduced by NOT relation or "is IN" type restriction
When some inserted RQL snippet generate more solutions than the original RQL,
the rewriter attempt to duplicate the snippet for each newly introduced
solution. There are though some cases where we do not want this behaviour in
case of ambiguities introduced by:

* NOT(X relation Y) expression, since it won't be
  equivalent to NOT(X relation Y1, Y1 is Type1) OR NOT(X relation Y2, Y2 is
  Type2) ;

* EXISTS(X relation Y, Y is IN (Type1, Type2) expression, since it's not
  actually necessary to split an explicitly introduced ambiguity (and it crash
  if we attempt to do so, so...).

In test, we've to modify the `rewrite()` function because in the newly
introduced test we need the same constraint to be applied to two variables in
the original query, and this was not supported before.

Notice the generated RQL in test is still *NOT CORRECT* "(EXISTS(NOT EXISTS() OR
EXISTS(...))", or at least isn't optimal. This will be fixed in a forthcoming
changeset.

Related to #17074119