[config] Avoid relying on shared_dir method when possible
From 3.26, some "data" directories (like web/data and i18n or wdoc) are
installed along with CubicWeb in site packages. So we can retrieve them
relatively to the module that needs them (namely "cubicweb" for "i18n"
and "cubicweb.web" for "web/data" and "wdoc"). Reason for doing this is
that "share_dir()" is not working when cubicweb is installed in
"develop" mode and we are about to drop this method.

Related to #17132990.

--HG--
branch : 3.26