[web/views] Fix AutoformSectionRelationTags.set_muledit_editable
It is a method call, so `self` is passed implicitely as first
argument, so it is useless to pass it also explicitely as second
argument. In fact, it does not work without this. A monkey patch
had been done in a private client's project, but the fix is
revelant for everyone.