Explicitly require repoze.lru for Pyramid tests
This used to be pulled by pyramid but the dependency got dropped in
pyramid version 1.10, see:

  https://docs.pylonsproject.org/projects/pyramid/en/latest/whatsnew-1.10.html#backward-incompatibilities

We already have an explicit dependency in setup.py, but it was missing
for tests.

Perhaps we could simply use '.[pyramid]' in this requirements file? Not
sure what side effect it'd have though...

--HG--
branch : 3.26